Switch to jnp.float16

+++ src/utils/config.py
@@ -6,7 +6,7 @@ def safe_array(*args, **kwargs):
     if 'dtype' not in kwargs and args and isinstance(args[0], (list, tuple, jnp.ndarray)):
         first_elem = args[0][0] if len(args[0]) > 0 else None
         if isinstance(first_elem, (float, int)):
-            kwargs['dtype'] = jnp.float32
+            kwargs['dtype'] = jnp.float16
     return _old_array(*args, **kwargs)
 
 jnp.array = safe_array
